+++ uikit-custom-sample/src/main/java/com/sendbird/uikit/customsample/groupchannel/activities/CustomChannelActivity.java
@@ -40,6 +40,7 @@ public class CustomChannelActivity extends ChannelActivity {
                 .setInputLeftButtonListener(v -> showMessageTypeDialog())
                 .setMessageListParams(null)
                 .setUseMessageGroupUI(useMessageGroupUI)
+                .showInputRightButtonAlways()
                 .build();
     }